Change to only display HSV color chooser

+++ src/de/devmil/common/ui/color/HsvAlphaSelectorView.java
@@ -166,7 +166,7 @@ public class HsvAlphaSelectorView extends LinearLayout {
 		placeSelector();
 	}
 	
-	public int getAlpha()
+	public float getAlpha()
 	{
 		return alpha;
 	}
